requirements.txt
absl-py==1.4.0
accelerate @ git+https://github.com/huggingface/accelerate.git@40007b4e977d07f00bb31f768425d17c4bcbf08c
aiohttp==3.9.5
aiosignal==1.3.1
ale-py==0.8.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array-record==0.4.1
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
AutoROM==0.6.1
AutoROM.accept-rom-license==0.6.1
Babel==2.14.0
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
bleach==6.1.0
blinker==1.8.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
Cython==3.0.8
data-setup==0.0.5
dataclasses==0.6
dataclasses-json==0.6.7
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.8
docutils==0.20.1
easyocr==1.7.2
entrypoints==0.4
et-xmlfile==1.1.0
etils==1.5.2
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastjsonschema==2.19.1
filelock==3.13.2
Flask==3.0.3
flatbuffers==24.3.25
flax==0.7.5
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.4.0
gin-config==0.5.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.15.0
google-api-python-client==2.114.0
google-auth==2.26.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.62.0
graphviz==0.20.3
greenlet==3.0.3
grpcio==1.60.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
httplib2==0.22.0
huggingface-hub==0.23.3
idna==3.6
imageio==2.35.1
immutabledict==4.1.0
import-ipynb==0.1.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.29.0
ipynb-py-convert==0.4.6
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.3.0
jax==0.4.23
jaxlib==0.4.23
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupytext==1.16.2
kaggle==1.6.3
keras==2.11.0
Keras-Preprocessing==1.1.2
keyring==24.3.0
kiwisolver==1.4.5
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-text-splitters==0.2.2
langsmith==0.1.93
lazy_loader==0.4
libclang==16.0.6
lvis==0.5.3
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mypy-extensions==1.0.0
namex==0.0.8
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.2.1
nh3==0.2.15
ninja==1.11.1.1
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
object-detection==0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
opt-einsum==3.3.0
optax==0.1.8
optree==0.12.1
orbax-checkpoint==0.4.4
orjson==3.10.6
overrides==7.4.0
p2j==1.3.2
packaging==23.2
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
peft @ git+https://github.com/huggingface/peft.git@0b40d1a304fc4d7d10baef14f913baff533dc2dc
pillow==10.2.0
pkginfo==1.9.6
platformdirs==4.1.0
portalocker==2.8.2
prometheus-client==0.19.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==3.19.6
psutil==5.9.7
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pygame==2.6.0
pyglet==2.0.17
Pygments==2.17.2
pyj==0.0
Pympler==1.1
pyparsing==3.1.1
PyPDF2==3.0.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-bidi==0.6.0
python-dateutil==2.8.2
python-docx==1.1.2
python-json-logger==2.0.7
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==5.4.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
readme-renderer==42.0
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
sacrebleu==2.2.0
safetensors==0.4.2
scikit-image==0.24.0
scikit-learn==1.3.2
scipy==1.11.4
semver==3.0.2
Send2Trash==1.8.2
sentencepiece==0.1.99
seqeval==1.2.2
shapely==2.0.6
Shimmy==1.3.0
six==1.16.0
slim==0.1
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.31
stable_baselines3==2.3.2
stack-data==0.6.3
streamlit==1.12.0
summarizer==0.0.7
sympy==1.12
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2.2
tensorflow==2.11.1
tensorflow-addons==0.22.0
tensorflow-datasets==4.9.0
tensorflow-decision-forests==1.8.1
tensorflow-estimator==2.11.0
tensorflow-hub==0.15.0
tensorflow-intel==2.11.1
tensorflow-io==0.31.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.0
tensorflow-model-optimization==0.7.5
tensorflow-object-detection-api==0.1.1
tensorflow-text==2.10.0
tensorflowjs==4.16.0
tensorstore==0.1.45
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tf-models-official==2.10.1
tf-slim==1.1.0
Theano==0.9.0
threadpoolctl==3.2.0
tifffile==2024.8.30
tiktoken==0.7.0
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.2.1
torch-tb-profiler==0.4.3
torchinfo==1.8.0
torchvision==0.17.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers @ git+https://github.com/huggingface/transformers.git@25245ec26dc29bcf6102e1b4ddd0dfd02e720cf5
twine==4.0.2
typeguard==2.13.3
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.1.0
validators==0.33.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wrapt==1.14.1
wurlitzer==3.0.3
yarl==1.9.4
zipp==3.17.0
